auduting, rate, homepage ishladi

+++ src/main/java/uz/pdp/heymasterapp/repository/RateRepository.java
@@ -9,6 +9,6 @@ import uz.pdp.heymasterapp.entity.enums.RoleEnum;
 
 import java.util.List;
 
-public interface RateRepository extends JpaRepository<Rate,Integer> {
+public interface RateRepository extends JpaRepository<Rate,Long> {
 
 }
